VAS bug 15790: fix '_init: not found' error
Clean up the way that the platform/style code works.
Also, add a -t option (not documented) to allow testing of the hook.
Used sed and BEGIN/END tags instead of head_minus_lines in the generic
shell fragments, since this will tie the hooks to the version of dnsupdate
and avoid removal of code blind to other modifications by the admin.